Review submission 8
Some dumbass forgot to go to next entry in his query

It's me. I'm the dumbass.

I'm feeling a whole lot of deja vu right now. But whatever, everything works now.

Oh, I forgot, I slightly reworked the UX so that when the user is logged in, the intro stuff wouldn't show.