Set up link for reviewer sign up plus some complications
So I set up the links for reviewer sign up, which led me to discover that the login doesn't set the reviewer status properly (again another case of missing = ? in my query), *then* I discovered that the <check> in my dynamically added article list wouldn't work. I worked around it by adding the check to the modal, since that's attached to the document instead of the script. Added in a little bit of fluff lol