Hosting pains 2 electric boogaloo
So I switched to ngrox from a friend's recommendation, and... it's not as terribly broken as before? It's not like pagekite because at least it actually works instead of gettting a 404, and it works better than x10hosting because scripts are actually loading now. But I'm getting this warning:

Cross-Origin Request Blocked: The Same Origin Policy disallows reading the remote resource at https://localhost:8088/. (Reason: CORS request did not succeed).

And because of this, most of my ajax calls return an error but work properly.

Also I couldn't have a customizable url. It gives me random url unless I sign up for their monthly service. And also since I'm basically just exposing my localhost, gotta keep my laptop online as often as possible.

Also added an SSL certificate. It didn't do anything but it didn't harm anything so it stays there for now until I make sure if it works fine or not.